renderer - batch generate qrcards Fixes #37
