+++ src/preload.ts
@@ -1,6 +1,7 @@
 import { contextBridge, ipcRenderer } from "electron";
 import { QRCodeOptions } from "qrcode";
-import { IVcardForm } from "./formToVcard";
+import { IImgFileDesc, IVcardForm } from "./types";
+
 
 contextBridge.exposeInMainWorld('qrapi', {
     qrcode: (data: string, o?: QRCodeOptions) => ipcRenderer.invoke('qrcode', data, o),
@@ -8,6 +9,9 @@ contextBridge.exposeInMainWorld('qrapi', {
 });
 contextBridge.exposeInMainWorld('vcardapi', {
     vcard: (formObj: IVcardForm) => ipcRenderer.invoke('vcard', formObj),
+});
+contextBridge.exposeInMainWorld('zipapi', {
+    zipimages: (images: IImgFileDesc[]) => ipcRenderer.invoke('zipimages', images),
 });
 
 // window.addEventListener('DOMContentLoaded', () => {
